+++ lib/capistrano/fiftyfive.rb
@@ -1,6 +1,0 @@
-require "capistrano/mb"
-
-compatibility_warning(
-  "The capistrano-fiftyfive gem has been renamed to capistrano-mb. "\
-  'Use `require "capistrano/mb"` to ensure compatibility with future versions.'
-)